// frontend/src/app/api/model/pagedResponseBookshelfSummaryDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BookshelfSummaryDto } from './bookshelfSummaryDto';


export interface PagedResponseBookshelfSummaryDto { 
    content?: Array<BookshelfSummaryDto>;
    pageNumber?: number;
    pageSize?: number;
    totalElements?: number;
    totalPages?: number;
    isLast?: boolean;
}