// frontend/src/app/api/model/reviewViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReviewViewDto { 
    id?: string;
    title?: string;
    comment?: string;
    rating?: number;
    userId?: string;
    username?: string;
    targetId?: string;
    targetType?: string;
    createdAt?: string;
    updatedAt?: string;
}